JPERF-181: Include integration tests in the main `build` task tree
Instead of testing acceptance only on release time or as opt-in on branches,
let's test the critical features for every pull request.
This way:
* we'll have more confidence in the pull requests
* there's less need to operate on snapshots
* we fail fast on merge time rather than at release-time, avoiding reverts
It comes at a cost of longer test times and higher barrier of entry due to AWS setup and run costs,
but it makes sense to expect AWS credentials for an AWS-centric module.