JPERF-945: Reproduce the Kotlin metadata bug
Rewrite locks: `./gradlew dep --write-locks`.
As you can see, the newest `aws-resources` in the declared range _is_ `1.10.1`.
The "is" means "as of the time of this commit",
so you might have a different result if you repeat it later on.
Luckily dependency locks let us freeze this moment in time.

To see the broken compatibility, run `./gradlew compileTestKotlin`.