Glue DC node tasks together
The first phase of `jira-node-X` tasks was merely building objects,
without actual provisioning, so it was instant.
Previously the `provision jira-node-X` task was happening after
the `provision load balancer` task, but those tasks are independent.
They can run in parallel, so `provision jira-node-X` can be shifted
earlier. Now it's called `install jira-node-X` and publishes SSH and EC2
instance details as events as early as possible.