Refactor `runLoadBalancerTest` a little
The routeIds size assertion failed on CI:
```
org.junit.ComparisonFailure: expected:<[2]> but was:<[1]>
```
Now we should get a bit better error message.
We will see the singular route and its count.
We might see that it's not the full 1000.
Or it might just show a single route with 1000 count.

We'll also get another CI trigger and see if it was a flake.
It' s possible that the problem existed for a longer time,
but `OracleJDK` flakiness was covering it up.